fix: venv symlink already exist when installing again into venv (#212)

+++ crates/rattler_installs_packages/src/python_env/venv.rs
@@ -272,7 +272,9 @@ prompt = {}"#,
 
             for bin_name in python_bins.into_iter() {
                 let venv_python_bin = venv_bin.join(bin_name);
-                if !venv_python_bin.exists() && venv_exe_path != venv_python_bin {
+                if (!venv_python_bin.exists() && !venv_python_bin.is_symlink())
+                    && venv_exe_path != venv_python_bin
+                {
                     copy_file(venv_exe_path, &venv_python_bin)?;
                 }
             }
